Add/update reusable empty components and improve history/profile handling - create AboutEmpty component for unauthenticated profile state - Update spacing for GameHistoryEmpty component - move empty states handling to corresponding pages (profile/history) - improve layout consistency between Home and Profile pages
