+++ frontend/src/pages/History.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useOutletContext } from "react-router"
-import { Spinner } from "@/components/ui/Spinner"
 import { userService } from "@/services/userService"
 import type { Match } from "@/lib/match"
 import { GameHistoryCard } from "@/components/home/GameHistoryCard"
+import { GameHistoryEmpty } from "@/components/home/GameHistoryEmpty";
 
 interface User {
 	id: number
@@ -45,9 +45,9 @@ export default function History() {
 
   	if (!user || loading) {
     	return (
-    		<div className="flex justify-center mt-20">
-        		<Spinner variant="cyan" size="lg" />
-      		</div>
+    		<section className="w-full max-w-3xl mx-auto px-6 py-10 text-white">
+          		<GameHistoryEmpty/>
+      		</section>
     	)
   	}
 
@@ -60,7 +60,10 @@ export default function History() {
 		</h1>
 
 		{/* CARD */}
-        <GameHistoryCard matches={matches} />
+        <GameHistoryCard
+		matches={matches}
+		user={user}
+		/>
 
 
     </section>
